Emacs: copy keybindings, hooks from Prelude to Spacemacs.
Copy keybindings for window movement, Paredit, Ex mode, and Cider REPL
from Prelude (in .emacs.d.backup) to Spacemacs (in .spacemacs).

NOTE: I want to add some functions to hooks and also run them from
functions added to other hooks. It seems to work to set a symbol, hook-sym,
to the function-name symbol, fn-sym, and then
*   Use (run-hooks 'hook-sym) in the other functions that I want to mix
    fn-sym into.
*   Use (add-hook '<hook> (lambda () (run-hooks 'hook-sym))) to add fn-sym
    to <hook>.
Is this the only or best way? I don't know.