nvim: Scope finding to git repos
I typically want to consider a git repo to be a project and
project_nvim_project.find_pattern_root seems to find the nearest
containing git repo so I use that function.

I first tried using project_nvim_project.get_project_root but that
detects a project whenever an LSP server detects a project. For example,
it detects each heuristics plugin as a project. That's great for LSP
operations but it's not what I want for finding files.

If I want to narrow searches, I think telescope provides a facility for
that.