nvim: Reimplement some LSP interactions to accept callback
I want to be able to compose LSP operations with editor operations (and,
really, with anything else).  The low-level LSP interactions are
requests and responses that are either
* Asynchronous and accept callback functions, or
* Synchronous and return results

However, the functionality above that, up to the user-facing functions,
are fire-and-forget, which doesn't compose.

As such, I'm re-implementing the relevant functionality on top of the
low-level LSP interactions.