nvim: Add aerial document symbols outline plugin
On the plus side, aerial allows a per-window document-symbols outline.

On the minus side, I have it set to show the document symbols outline
when I open a file. However, if I split the window horizontally, then
the outline of the original window (itself a window) remains full-height
while the new split itself opens an outline inside of its space.

So we go from
+--+--------+
|  |        |
|  |        |
|  |        |
+--+--------+
to
+--+--------+
|  |        |
|  +--+-----|
|  |  |     |
+--+--------+

when what we actually want is
+--+--------+
|  |        |
+--+--------|
|  |        |
+--+--------+

Similarly, since the outline is itself a window, then we can navigate to
it with the window-movement commands but we also have to navigate
through it when we have vertical splits and each has its own outline
window.

We could correct both of these things but it may be involved.