nvim: Use local fork of aerial plugin
Aerial can operate in global mode (one outline, displays symbols for the
active window) or in per-window mode (one outline per window).

It also has auto-open behavior that triggers on various events but which
doesn't re-open the outline if the user has manually closed it.

However, that auto-open logic is written in the context of the global
mode. I'm reimplementing it to work in the context of both global and
per-window modes.