nvim: Skip aerial outline windows when moving between windows
As in the code comment, this works but it's really dumb. Also, my
at-the-edge heuristic may not always hold. In fact, it is at least
dependent on the aerial window being displayed to the left of the
associated window.

Idea: Can we instead get the window layout, determine the window to
which to move, and then actually move the cursor?