emacs: Replace Doom's undo module with a la carte vundo
Emacs 28 added
* undo-only as a proper undo
* undo-redo as a proper redo
It seems to me these additions make undo-fu obsolete.
Vundo is an undo tree that works with the built-in emacs undo
mechanisms. The Doom undo module uses Vundo but it also uses undo-fu.

Let's build on this general idea: carve away at the excess that Doom
provides :D