emacs: Use consult like telescope.nvim to find files
Search for files with `fd' in DIR by using fd to (synchronously) list
all files in the directory and then filter the results using
completing read (which I have using fzf).

This is like telescope.nvim (though this is synchronous) and unlike
the built-in `consult-fd` behavior, which uses `fd` for primary
filtering and `completing-read` for secondary filtering.
The latter approach is more powerful but this approach is what I want
most of the time.